etl-gcp-json-bigquery/requirements.txt
pandas==1.5.3
google-cloud-bigquery==3.4.2
google-cloud-storage==2.7.0
